chore(rcm): cyclomatic complexity in remote config client
