+++ ddtrace/internal/remoteconfig/client.py
@@ -145,43 +145,8 @@ def _load_json(data):
     return json.loads(data)
 
 
-def _extract_target_file(payload, target, config):
-    # type: (AgentPayload, str, ConfigMetadata) -> Optional[Mapping[str, Any]]
-    candidates = [item.raw for item in payload.target_files if item.path == target]
-    if len(candidates) != 1 or candidates[0] is None:
-        log.debug("invalid target_files for %r", target)
-        return None
-
-    try:
-        raw = base64.b64decode(candidates[0])
-    except Exception:
-        raise RemoteConfigError("invalid base64 target_files for {!r}".format(target))
-
-    computed_hash = hashlib.sha256(raw).hexdigest()
-    if computed_hash != config.sha256_hash:
-        raise RemoteConfigError(
-            "mismatch between target {!r} hashes {!r} != {!r}".format(target, computed_hash, config.sha256_hash)
-        )
-
-    try:
-        return _load_json(raw)
-    except Exception:
-        raise RemoteConfigError("invalid JSON content for target {!r}".format(target))
-
-
-def _parse_target(target, metadata):
-    # type: (str, TargetDesc) -> ConfigMetadata
-    m = TARGET_FORMAT.match(target)
-    if m is None:
-        raise RemoteConfigError("unexpected target format {!r}".format(target))
-    _, product_name, config_id, _ = m.groups()
-    return ConfigMetadata(
-        id=config_id,
-        product_name=product_name,
-        sha256_hash=metadata.hashes.get("sha256"),
-        length=metadata.length,
-        tuf_version=metadata.custom.get("v"),
-    )
+AppliedConfigType = Dict[str, ConfigMetadata]
+TargetsType = Dict[str, ConfigMetadata]
 
 
 class RemoteConfigClient(object):
@@ -222,7 +187,7 @@ class RemoteConfigClient(object):
             app_version=ddtrace.config.version,
             tags=[":".join(_) for _ in tags.items()],
         )
-        self.cached_target_files = []  # type: List[Dict[str, Any]]
+        self.cached_target_files = []  # type: List[AppliedConfigType]
         self.converter = cattr.Converter()
 
         # cattrs doesn't implement datetime converter in Py27, we should register
@@ -240,7 +205,7 @@ class RemoteConfigClient(object):
         self.converter.register_structure_hook(SignedTargets, base64_to_struct)
 
         self._products = dict()  # type: MutableMapping[str, ProductCallback]
-        self._applied_configs = dict()  # type: Mapping[str, ConfigMetadata]
+        self._applied_configs = dict()  # type: AppliedConfigType
         self._last_targets_version = 0
         self._last_error = None  # type: Optional[str]
         self._backend_state = None  # type: Optional[str]
@@ -278,6 +243,45 @@ class RemoteConfigClient(object):
 
         return json.loads(data)
 
+    @staticmethod
+    def _extract_target_file(payload, target, config):
+        # type: (AgentPayload, str, ConfigMetadata) -> Optional[Mapping[str, Any]]
+        candidates = [item.raw for item in payload.target_files if item.path == target]
+        if len(candidates) != 1 or candidates[0] is None:
+            log.debug("invalid target_files for %r", target)
+            return None
+
+        try:
+            raw = base64.b64decode(candidates[0])
+        except Exception:
+            raise RemoteConfigError("invalid base64 target_files for {!r}".format(target))
+
+        computed_hash = hashlib.sha256(raw).hexdigest()
+        if computed_hash != config.sha256_hash:
+            raise RemoteConfigError(
+                "mismatch between target {!r} hashes {!r} != {!r}".format(target, computed_hash, config.sha256_hash)
+            )
+
+        try:
+            return _load_json(raw)
+        except Exception:
+            raise RemoteConfigError("invalid JSON content for target {!r}".format(target))
+
+    @staticmethod
+    def _parse_target(target, metadata):
+        # type: (str, TargetDesc) -> ConfigMetadata
+        m = TARGET_FORMAT.match(target)
+        if m is None:
+            raise RemoteConfigError("unexpected target format {!r}".format(target))
+        _, product_name, config_id, _ = m.groups()
+        return ConfigMetadata(
+            id=config_id,
+            product_name=product_name,
+            sha256_hash=metadata.hashes.get("sha256"),
+            length=metadata.length,
+            tuf_version=metadata.custom.get("v"),
+        )
+
     def _build_payload(self, state):
         # type: (Mapping[str, Any]) -> Mapping[str, Any]
         return dict(
@@ -311,59 +315,24 @@ class RemoteConfigClient(object):
         return state
 
     def _process_targets(self, payload):
-        # type: (AgentPayload) -> Tuple[Optional[int], Optional[str], Optional[Mapping[str, ConfigMetadata]]]
+        # type: (AgentPayload) -> Tuple[Optional[int], Optional[str], Optional[TargetsType]]
         if payload.targets is None:
             # no targets received
             return None, None, None
 
         signed = payload.targets.signed
-        targets = dict()
+        targets = dict()  # type: TargetsType
 
         for target, metadata in signed.targets.items():
-            config = _parse_target(target, metadata)
+            config = self._parse_target(target, metadata)
             if config is not None:
                 targets[target] = config
 
         backend_state = signed.custom.get("opaque_backend_state")
         return signed.version, backend_state, targets
 
-    def _process_response(self, data):
-        # type: (Mapping[str, Any]) -> None
-        try:
-            # log.debug("response payload: %r", data)
-            payload = self.converter.structure_attrs_fromdict(data, AgentPayload)
-        except Exception:
-            log.debug("invalid agent payload received: %r", data, exc_info=True)
-            raise RemoteConfigError("invalid agent payload received")
-
-        paths = {_.path for _ in payload.target_files}
-        paths = paths.union({_["path"] for _ in self.cached_target_files})
-
-        # !(payload.client_configs is a subset of paths or payload.client_configs is equal to paths)
-        if not set(payload.client_configs) <= paths:
-            raise RemoteConfigError("Not all client configurations have target files")
-
-        # 1. Deserialize targets
-        last_targets_version, backend_state, targets = self._process_targets(payload)
-        if last_targets_version is None or targets is None:
-            log.debug("No targets in configuration payload")
-            for callback in self._products.values():
-                callback(None, None)
-            return
-
-        client_configs = {k: v for k, v in targets.items() if k in payload.client_configs}
-        log.debug("Retrieved client configs last version %s: %s", last_targets_version, client_configs)
-
-        for target in payload.target_files:
-            if (payload.targets.signed.targets and not payload.targets.signed.targets.get(target.path)) and (
-                client_configs and not client_configs.get(target.path)
-            ):
-                raise RemoteConfigError(
-                    "target file %s not exists in client_config and signed targets" % (target.path,)
-                )
-
-        # 2. Remove previously applied configurations
-        applied_configs = dict()
+    def _remove_previously_applied_configurations(self, applied_configs, client_configs, targets):
+        # type: (AppliedConfigType, TargetsType, TargetsType) -> None
         for target, config in self._applied_configs.items():
             callback_action = None
             if target in client_configs and targets.get(target) == config:
@@ -382,7 +351,8 @@ class RemoteConfigClient(object):
                 log.debug("error while removing product %s config %r", config.product_name, config)
                 continue
 
-        # 3. Load new configurations
+    def _load_new_configurations(self, applied_configs, client_configs, payload):
+        # type: (AppliedConfigType, TargetsType, AgentPayload) -> None
         for target, config in client_configs.items():
             callback = self._products[config.product_name]
 
@@ -390,12 +360,12 @@ class RemoteConfigClient(object):
             if applied_config == config:
                 continue
 
-            config_content = _extract_target_file(payload, target, config)
+            config_content = self._extract_target_file(payload, target, config)
             if config_content is None:
                 continue
 
             try:
-                log.debug("Load new configuration: %s. content %s", target, config_content)
+                log.debug("Load new configuration: %s. content ", target)
                 callback(config, config_content)
             except Exception:
                 log.debug("error while loading product %s config %r", config.product_name, config)
@@ -403,10 +373,7 @@ class RemoteConfigClient(object):
             else:
                 applied_configs[target] = config
 
-        self._last_targets_version = last_targets_version
-        self._applied_configs = applied_configs
-        self._backend_state = backend_state
-
+    def _add_apply_config_to_cache(self):
         if self._applied_configs:
             cached_data = []
             for target, config in self._applied_configs.items():
@@ -421,12 +388,68 @@ class RemoteConfigClient(object):
         else:
             self.cached_target_files = []
 
+    def _validate_config_exists_in_target_paths(self, payload_client_configs, payload_target_files):
+        # type: (Set[str], List[TargetFile]) -> None
+        paths = {_.path for _ in payload_target_files}
+        paths = paths.union({_["path"] for _ in self.cached_target_files})
+
+        # !(payload.client_configs is a subset of paths or payload.client_configs is equal to paths)
+        if not set(payload_client_configs) <= paths:
+            raise RemoteConfigError("Not all client configurations have target files")
+
+    @staticmethod
+    def _validate_signed_target_files(payload_target_files, payload_targets_signed, client_configs):
+        # type: (List[TargetFile], Targets, TargetsType) -> None
+        for target in payload_target_files:
+            if (payload_targets_signed.targets and not payload_targets_signed.targets.get(target.path)) and (
+                client_configs and not client_configs.get(target.path)
+            ):
+                raise RemoteConfigError(
+                    "target file %s not exists in client_config and signed targets" % (target.path,)
+                )
+
+    def _process_response(self, data):
+        # type: (Mapping[str, Any]) -> None
+        try:
+            # log.debug("response payload: %r", data)
+            payload = self.converter.structure_attrs_fromdict(data, AgentPayload)
+        except Exception:
+            log.debug("invalid agent payload received: %r", data, exc_info=True)
+            raise RemoteConfigError("invalid agent payload received")
+
+        self._validate_config_exists_in_target_paths(payload.client_configs, payload.target_files)
+
+        # 1. Deserialize targets
+        last_targets_version, backend_state, targets = self._process_targets(payload)
+        if last_targets_version is None or targets is None:
+            log.debug("No targets in configuration payload")
+            for callback in self._products.values():
+                callback(None, None)
+            return
+
+        client_configs = {k: v for k, v in targets.items() if k in payload.client_configs}
+        log.debug("Retrieved client configs last version %s: %s", last_targets_version, client_configs)
+
+        self._validate_signed_target_files(payload.target_files, payload.targets.signed, client_configs)
+
+        # 2. Remove previously applied configurations
+        applied_configs = dict()  # type: AppliedConfigType
+        self._remove_previously_applied_configurations(applied_configs, client_configs, targets)
+
+        # 3. Load new configurations
+        self._load_new_configurations(applied_configs, client_configs, payload)
+
+        self._last_targets_version = last_targets_version
+        self._applied_configs = applied_configs
+        self._backend_state = backend_state
+
+        self._add_apply_config_to_cache()
+
     def request(self):
         # type: () -> None
         try:
             state = self._build_state()
             payload = json.dumps(self._build_payload(state))
-
             response = self._send_request(payload)
             if response is None:
                 return
